Plot background lines.

Plot background lines.

Usage

plot_background_lines(
  dat,
  line_values,
  line_se,
  plot_settings = plotsettings_lineplot()
)

Arguments

dat

Data.

line_values

Character vector with two elements. Column names in plot_dat[["plot_lines"]] containing the y-values for the plotted lines. Defaults to c("est_noTrendStart_noComp", "est_noTrendEnd_noComp"). If set to NULL, no lines will be plotted.

line_se

Character vector of the column name containing the standard errors for the plotted lines. Defaults to NULL, in which case they will be deducted from the line values.

plot_settings

Named list constructed with plotsettings_lineplot(). Defaults to a list with all settings set to 0. There are several predefined lists with optimized settings for different plots. See plotsettings_lineplot() for an overview.

Value

ggplot2 object.

Examples

# tbd