Plot a lineplot.

Plot a lineplot.

Usage

plot_lineplot(
  plot_dat,
  seperate_plot_var = "state_var",
  seperate_plot_var_order = NULL,
  seperate_plot_var_box = "wholeGroup",
  point_values = "est_noTrend_noComp",
  point_sig = "sig_noTrend_Comp_crossDiff_wholeGroup",
  line_values = c("est_noTrendStart_noComp", "est_noTrendEnd_noComp"),
  line_sig = "sig_Trend_noComp",
  label_est = "est_Trend_noComp",
  label_se = "se_Trend_noComp",
  label_sig_high = NULL,
  label_sig_bold = "sig_Trend_noComp",
  line_se = NULL,
  title_superscripts = NULL,
  years_lines = NULL,
  years_braces = NULL,
  plot_settings = plotsettings_lineplot()
)

Arguments

plot_dat

Input is a list prepared by prep_plot().`

seperate_plot_var

Character string of the column containing the tiles. For every unique value, a new tile will be plotted. Defaults to state_var.

seperate_plot_var_order

Character vector containing all unique elements in the seperate_plot_var column. The lineplot-tiles will be ordered according to the order of this vector.

seperate_plot_var_box

Character vector, containing strings from the seperate_plot_var-column, that should get a box drawn around them.

point_values

Character string of the column name in plot_dat[["plot_points"]] containing the y-values for the plotted points. Defaults to est_noTrend_noComp.

point_sig

Character string of the column name containing significance values for point_values. Defaults to "sig_noTrend_Comp_crossDiff_wholeGroup".

line_values

Character vector with two elements. Column names in plot_dat[["plot_lines"]] containing the y-values for the plotted lines. Defaults to c("est_noTrendStart_noComp", "est_noTrendEnd_noComp"). If set to NULL, no lines will be plotted.

line_sig

Character string of the column name containing significance values for line_values. Defaults to "sig_Trend_noComp", which will show the significance of the difference between two time points.

label_est

Character string of the column name containing the brace labels.

label_se

Character string of the column name containing the standard errors for label_est. Will be put in bracktes behind label_est.

label_sig_high

Character string of the column name containing significance values for label_est. Significant values will be marked by a raised 'a'. Normally, should be the comparison of the trend vs. the trend in whole Germany, which can be found in the trendDiff_cross parameter. Defaults to NULL, as this parameter is not always provided.

label_sig_bold

Character string of the column name containing significance values for label_est. Significant values will be marked as bold. Defaults to "sig_Trend_noComp".

line_se

Character vector of the column name containing the standard errors for the plotted lines. Defaults to NULL, in which case they will be deducted from the line values.

title_superscripts

Named list for superscripts at the plot_titles. The name of the list element has to be equal to the title, the value of the list element has to be the superscript. Defaults to NULL.

years_lines

List of numeric vectors containing the start and end year, between which a trend line should be plotted. Per default, lines are drawn from every year to the next consecutive year.

years_braces

List of numeric vectors containing the start and end year, between which a brace should be plotted. Per default, braces are drawn from the last year to every other year included in the data.

plot_settings

Named list constructed with plotsettings_lineplot(). Defaults to a list with all settings set to 0. There are several predefined lists with optimized settings for different plots. See plotsettings_lineplot() for an overview.

Value

ggplot2 object.

Examples

# tbd