Set parameters for the barplots.

Set parameters for the barplots.

Usage

plotsettings_tablebarplot(
  axis_x_lims = NULL,
  background_stripes_colour = NULL,
  bar_border_lines = NULL,
  bar_fill_colour = NULL,
  bar_frame_linetype = NULL,
  bar_label_nudge_x = NULL,
  bar_label_size = NULL,
  bar_line_size = NULL,
  bar_pattern_fill_colour = NULL,
  bar_pattern_type = NULL,
  bar_sig_type = NULL,
  bar_width = NULL,
  columns_alignment = NULL,
  columns_nudge_x = NULL,
  columns_table_sig_high_letter = NULL,
  columns_table_sig_high_letter_nudge_x = NULL,
  columns_width = NULL,
  headers_alignment = NULL,
  headers_background_colour = NULL,
  headers_background_width_x = NULL,
  headers_nudge_x = NULL,
  headers_nudge_y = NULL,
  font_size = NULL,
  pattern_spacing = NULL,
  pattern_width = NULL,
  default_list = NULL
)

Arguments

axis_x_lims

Numeric vector of length 2 for the x-axis limits. Will be set automatically, NULL (default).

background_stripes_colour

Named vector containing the two colours that should be used for the striped background. Defaults to NULL.

bar_border_lines

Logical indicating whether the barplot should receive dotted lines on its borders.

bar_fill_colour

Colour of the bar filling. Can be either one colour for all bars, or a (named) vector with the names of the groups specified in bar_fill. If no names are provided for the vector, the order of the factor levels of bar_fill will be used for determining the colour assignment.

bar_frame_linetype

Named vector with the bar frame linetypes. Names have to be found in the column defined in the bar_sig-argument ofplot_tablebar(). Defaults to solid.

bar_label_nudge_x

Numeric for nudging the bar labels in x direction.

bar_label_size

Numeric for the font size of the bar labels.

bar_line_size

Numeric for the line-size around the bar.

bar_pattern_fill_colour

Named vector with the filling colours for the bar pattern. Names of the vector must be found in the column specified in bar_pattern_fill. Defaults to white.

bar_pattern_type

Named vector with the pattern types for the barpattern.

bar_sig_type

Character string indicating whether levels of the grouping variable should be visualized by pattern fill ("pattern") or line type ("frame").

bar_width

Numeric between 0 and 1 specifying the width of the bar. Defaults to 0.4.

columns_alignment

Numeric vector with one element for each column, determining the text adjustement within the column. Can be 0 (left-aligned), 0.5 (central-aligned), or 1 (right-aligned). Defaults to 0.5.

columns_nudge_x

Numeric to nudge the column text in x direction. Defaults to 0.

columns_table_sig_high_letter

Character, that will be added on significant values defined by columns_table_sig_high.

columns_table_sig_high_letter_nudge_x

Numeric for nudging the superscript towards or away from a number.

columns_width

Numeric vector with relative column widths. Has to be equal to the number of columns (including the bar chart, if a bar chart is plotted) that are plotted in the table. Defaults to NULL, in which case all collumns will get the same width.

headers_alignment

Numeric vector with one element for each column, determining the text adjustement of the headers. Can be 0 (left-aligned), 0.5 (central-aligned), or 1 (right-aligned). Defaults to NULL, in which case the alignment of the columns will be adopted.

headers_background_colour

Colour of the background of the headers.

headers_background_width_x

Numeric for the width of the headers background. The defined value will be added to the top of the plot, to increase the area on the plots upper border. Defaults to 0.

headers_nudge_x

Numeric to nudge the column_headers in x direction. Defaults to 0.

headers_nudge_y

Numeric to nudge the column_headers in y direction. Defaults to 0.

font_size

Numeric vector with as many elements as columns for the font sizes of the columns. Defaults to 3.

pattern_spacing

Numeric for the gap between patterns.

pattern_width

Numeric for the width of the pattern stripes. Note that by default the pattern are the whit stripes, so an increase of the pattern_width parameter will increase the thickness of the white stripes.

default_list

Named list with predefined settings. Defaults to a list with all settings set to 0.

Value

A named list with settings for a table/barplot.

Examples

## Calling the function without any arguments will result in the default settings-list.
plotsettings_tablebarplot()
#> $axis_x_lims
#> NULL
#> 
#> $background_stripes_colour
#> [1] "white"
#> 
#> $bar_border_lines
#> [1] FALSE
#> 
#> $bar_fill_colour
#> [1] "white"
#> 
#> $bar_frame_linetype
#> [1] "solid"
#> 
#> $bar_label_size
#> [1] 2
#> 
#> $bar_label_nudge_x
#> [1] 0
#> 
#> $bar_line_size
#> [1] 0.5
#> 
#> $bar_pattern_fill_colour
#> [1] "white"
#> 
#> $bar_pattern_type
#> [1] "none"
#> 
#> $bar_sig_type
#> [1] "frame"
#> 
#> $bar_width
#> [1] 0.4
#> 
#> $columns_alignment
#> [1] 0.5
#> 
#> $columns_nudge_x
#> [1] 0
#> 
#> $columns_table_sig_high_letter
#> [1] "a"
#> 
#> $columns_table_sig_high_letter_nudge_x
#> [1] 0
#> 
#> $columns_width
#> NULL
#> 
#> $headers_alignment
#> NULL
#> 
#> $headers_background_colour
#> [1] "white"
#> 
#> $headers_background_width_x
#> [1] 0
#> 
#> $headers_nudge_x
#> [1] 0
#> 
#> $headers_nudge_y
#> [1] 0
#> 
#> $pattern_spacing
#> [1] 0.1
#> 
#> $pattern_width
#> [1] 0.5
#> 
#> $font_size
#> [1] 3
#> 

## Arguments will overvrite the respective list element:
plot_settings <- plotsettings_tablebarplot(font_size = 2.5)
plot_settings$font_size
#> [1] 2.5

## Colours can be defined like so:
plotsettings_tablebarplot(
  bar_fill_colour = c("blue",
                      grDevices::rgb(10, 13, 82, maxColorValue = 255)
                      )
)
#> $axis_x_lims
#> NULL
#> 
#> $background_stripes_colour
#> [1] "white"
#> 
#> $bar_border_lines
#> [1] FALSE
#> 
#> $bar_fill_colour
#> [1] "blue"    "#0A0D52"
#> 
#> $bar_frame_linetype
#> [1] "solid"
#> 
#> $bar_label_size
#> [1] 2
#> 
#> $bar_label_nudge_x
#> [1] 0
#> 
#> $bar_line_size
#> [1] 0.5
#> 
#> $bar_pattern_fill_colour
#> [1] "white"
#> 
#> $bar_pattern_type
#> [1] "none"
#> 
#> $bar_sig_type
#> [1] "frame"
#> 
#> $bar_width
#> [1] 0.4
#> 
#> $columns_alignment
#> [1] 0.5
#> 
#> $columns_nudge_x
#> [1] 0
#> 
#> $columns_table_sig_high_letter
#> [1] "a"
#> 
#> $columns_table_sig_high_letter_nudge_x
#> [1] 0
#> 
#> $columns_width
#> NULL
#> 
#> $headers_alignment
#> NULL
#> 
#> $headers_background_colour
#> [1] "white"
#> 
#> $headers_background_width_x
#> [1] 0
#> 
#> $headers_nudge_x
#> [1] 0
#> 
#> $headers_nudge_y
#> [1] 0
#> 
#> $pattern_spacing
#> [1] 0.1
#> 
#> $pattern_width
#> [1] 0.5
#> 
#> $font_size
#> [1] 3
#> 

## Or, to get better control over the colour assignment
## we can also directly name the colours with the different groups
## defined in the bar_sig-argument of plot_tablebar()
## (as these groups are used to distinguish between different bar colours):
plotsettings_tablebarplot(
 bar_fill_colour = c(fill_group_1 = "blue",
                     fill_group_2 = grDevices::rgb(
                       10, 13, 82, maxColorValue = 255)
                     )
)
#> $axis_x_lims
#> NULL
#> 
#> $background_stripes_colour
#> [1] "white"
#> 
#> $bar_border_lines
#> [1] FALSE
#> 
#> $bar_fill_colour
#> fill_group_1 fill_group_2 
#>       "blue"    "#0A0D52" 
#> 
#> $bar_frame_linetype
#> [1] "solid"
#> 
#> $bar_label_size
#> [1] 2
#> 
#> $bar_label_nudge_x
#> [1] 0
#> 
#> $bar_line_size
#> [1] 0.5
#> 
#> $bar_pattern_fill_colour
#> [1] "white"
#> 
#> $bar_pattern_type
#> [1] "none"
#> 
#> $bar_sig_type
#> [1] "frame"
#> 
#> $bar_width
#> [1] 0.4
#> 
#> $columns_alignment
#> [1] 0.5
#> 
#> $columns_nudge_x
#> [1] 0
#> 
#> $columns_table_sig_high_letter
#> [1] "a"
#> 
#> $columns_table_sig_high_letter_nudge_x
#> [1] 0
#> 
#> $columns_width
#> NULL
#> 
#> $headers_alignment
#> NULL
#> 
#> $headers_background_colour
#> [1] "white"
#> 
#> $headers_background_width_x
#> [1] 0
#> 
#> $headers_nudge_x
#> [1] 0
#> 
#> $headers_nudge_y
#> [1] 0
#> 
#> $pattern_spacing
#> [1] 0.1
#> 
#> $pattern_width
#> [1] 0.5
#> 
#> $font_size
#> [1] 3
#>