Prepare data for plotting a barplot.

Prepare data for plotting a barplot.

Usage

prep_no_trend(
  dat,
  columns,
  competence,
  grouping_var = "",
  group_var = "group",
  state_var = "TR_BUNDESLAND",
  competence_var = "kb",
  states = NULL,
  sig_niveau = 0.05,
  parameter = "mean"
)

Arguments

dat

Input data.frame stemming from eatRep.

columns

Character string containing the column in data that contains the column names for the table_plot.

competence

Character string containing the competence that should be plotted.

grouping_var

Character string containing the column name in dat that should be used to distinguish between subgroups.

group_var

Character string containing the column name in dat that contains the different group memberships in one string. Defaults to "group".

state_var

Character string containing the column name in dat that should be used to distinguish between groups that should be plotted seperatly. Normally, this should be the states ("Bundesländer"). Therfore, defaults to "TR_BUNDESLAND".

competence_var

Character string containing the column name in dat that contains the different competences. Defaults to "kb".

states

Character vector of the states that should be plotted.

sig_niveau

Numeric indicating the border, below which p-values will be considered significant. Defaults to 0.05.

parameter

Character string of the parameter value that should be reported. Defaults to "mean".

Value

Returns a data.frame which can be used as input for plot_bar.

Examples

# tbd