Plot braces below plot.

Plot braces below plot.

Usage

plot_braces(
  dat,
  y_range,
  label_est = NULL,
  label_se = NULL,
  label_sig_high = NULL,
  label_sig_bold = NULL,
  plot_settings = plotsettings_lineplot()
)

Arguments

dat

Prepared Trend data.

y_range

Numeric vector with two elements, indicating the min and max for the y axis.

label_est

Character string of the column name containing the brace labels.

label_se

Character string of the column name containing the standard errors for label_est. Will be put in bracktes behind label_est.

label_sig_high

Character string of the column name containing significance values for label_est. Significant values will be marked by a raised 'a'. Normally, should be the comparison of the trend vs. the trend in whole Germany, which can be found in the trendDiff_cross parameter. Defaults to NULL, as this parameter is not always provided.

label_sig_bold

Character string of the column name containing significance values for label_est. Significant values will be marked as bold. Defaults to "sig_Trend_noComp".

plot_settings

Named list constructed with plotsettings_lineplot(). Defaults to a list with all settings set to 0. There are several predefined lists with optimized settings for different plots. See plotsettings_lineplot() for an overview.

Value

ggplot2 object.

Examples

##